Add support for multilingual bibliographic entries.
Below is the short story behind choosing packages `biblatex' and
`polyglossia' over `babelbib' and 'babel', switching to XeTeX
and applying a specific font family.

The initial plan was to use the package `babelbib' which allows
to specify language on a per-entry basis as well as the entry
format using its own intrinsic bibliographic styles (e.g.,
babplain).

However, an author-year style for the citation would be preferred
over the default (ID). In this case, package `natbib' would allow
to specify the format. Unfortunately, it would also require to use
its own bibliographic styles (e.g., plainnat).

Thus, the two packages would not work together (since both require
to set a different option to the \bibliographystyle command).

An alternative to these would be the (newer) package `biblatex'
which provides all the aforementioned capabilities. However, this
particular package is incompatible with the `ucs' package
which is automatically loaded when specifying the non-standard
utf8x option for the input encoding. The latter permitted non-latin
text to be directly input into the document when using the `babel'
package.

To tackle this, a Unicode-enabled engine, XeTeX, which expects
UTF-8 input encoding by default (as suggested in its
documentation: Incompatible Packages / ucs), in conjunction with
the package `polyglossia' were selected.

A requirement of `polyglossia' is to specify otf/ttf fonts for use
with the selected non-english languages (using commands from the
package `fontspec'). In order to meet this requirement but, still
maintain the overall appearance of the default fonts used by LaTeX,
the Computer Modern Unicode (cm-unicode) family is chosen. This
particular font family is provided by TeXLive and is under the X11
License.