Add TCRT section and first few paragraphs.
Includes subsection `Coupling factor' and subsubsections
`Reflecting surfaces' and 'Working distance'. Images are included
only under Operational distance. It should be noted, though, that
if multiple figures are to be placed one next to the other, the
more appropriate `subcaption' package should be preferred.

This commit adds the appropriate image source references and a
`nottype' filtering on images from the global \printbibliography
command. To that end, the obsolete `image' BibTeX entry type is
used.