Add the http_server module.
This is the central node of the HTTP server and lends itself code which was
temporarily spread over other files (such as `net' and `http_parser'). In
summary, it:

* Provides an HTTP request representation.
* Defines a set of HTTP tokens and macros for accessing them.
* Keeps track of the server host name and port.
* Registers exposed resources and corresponding handlers.

One secondary modification is in regard to the parameters of function
`stream_parse' (of module stream_util.h). Parameter `min' is no longer
specified. This is partly due to the way marcos are defined in `http_server'
and, consequently, how it is invoked by `http_parser'. Module `json_parser'
is also affected in that regard.