Add log search field validation and appropriate URI state update.
Now, whenever a valid search is initiated, the values are projected onto the
URI fragment. This, of course, results in stateLoad() to be triggered. This
way, the state handler of this page (log-page) can seamlessly operate with
searches triggered from within the UI as well as those inserted directly into
the address bar.